get profile and timeline
